solucion error controller cliente

+++ src/main/java/com/alura/hotelalura/springboottres/persitence/dto/reserva/MostrarReservacion.java
@@ -1,6 +1,6 @@
 package com.alura.hotelalura.springboottres.persitence.dto.reserva;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 public interface MostrarReservacion 
@@ -8,7 +8,7 @@ public interface MostrarReservacion
    String getReservaId();
    LocalDate getCheckIn();
    LocalDate getCheckOut();
-   BigDecimal getPrecioHabitacion();
+   Float getPrecioHabitacion();
    String getCategoria();
    Integer getNumeroHabitacion();
